Create standalone SecretPhraseModal
